requirements.txt
# Flask Backend Dependencies
flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.3.2
groq==0.9.0
httpx==0.27.2
gtts==2.5.0
deep-translator==1.11.4
Pillow==10.1.0
python-dotenv==1.0.0

# Production deployment
gunicorn==21.2.0
